+++ mayan/apps/sources/models/watch_folder_sources.py
@@ -3,8 +3,7 @@ from __future__ import unicode_literals
 import errno
 import fcntl
 import logging
-
-from pathlib2 import Path
+from pathlib import Path
 
 from django.db import models
 from django.utils.translation import ugettext_lazy as _